requirement.txt
asgiref==3.2.5
boto3==1.12.26
botocore==1.15.26
Django==3.0.4
django-cors-headers==3.2.1
django-mysql==3.3.0
django-s3-storage==0.13.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.9.5
jsonfield==3.1.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
